Day3Problem6
BONUS: Write a function that prints the least integer that is not present in a given list and cannot be represented by the summation of the sub-elements of the list. E. g. For a  = [1,2,5,7] the least integer not represented by the list is 4, and if a = [1,2,2,5,7] then the least non representable integer is 18.