Day5Problem3
Find maximum possible stolen value from houes.
There are n houses build in a line, each of which contains some value in it. A thief is going to steal the maximum value of these houses, but he can't steal in two adjacent houses because the owner of the stolen houses will tell his two neighbours left and right side. What is the maximum stolen value?